feat: キーワード抽出処理を別ファイルに分離
キーワード抽出ロジックを extractKeywords.ts に移動し、モジュールの構造を最適化。index.ts からキーワード抽出関数をインポートし、コードの可読性と保守性を向上。
